refact(pkg): move pkg out of versioning
since no version dependent methods are present, move pkg out of v1alpha1
